refactor: move data/selectors to data/apiHooks
and fix tests to mock useQuery.
